+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add entities cache for faster lookup of measurable data types
 - Use entities cache in measurement summary
 - Refactor: move fetch logic to cubit
+- Refresh query when typing in full-text search
 
 ## [0.8.227] - 2023-01-07
 ### Changed: